K8 Shell Driver/requirements.txt
cloudshell-automation-api>=8.2.0.0,<8.3.0.0
cloudshell-shell-core>=2.0.0,<2.1.0
jsonpickle==0.9.3
pycrypto==2.6.1
kubernetes